backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.13.1
asyncmy==0.2.11
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
scikit-learn>=1.5.2
pandas>=2.2.3
numpy>=2.1.0
joblib>=1.4.2
google-generativeai>=0.8.0
httpx>=0.27.0
slowapi>=0.1.9
aiofiles>=23.2.1
python-dateutil>=2.9.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
aiosqlite>=0.20.0
